requirements.txt
python-magic~=0.4.18
pdftotext~=2.1.5
tabulate~=0.8.7